Fix typo in self._chinese_mode variable
When porting to GObjectIntrospection, I made a typing mistake
in the self._chinese_mode variable. This caused a problem when
the dconf key for “chinesemode” was not yet set. Then the property
for selecting the Chinese mode in the ibus menu was empty and
and selecting the empty line with the mouse just caused an error.

This commit fixes this.