Preedit needs to be updated on page-up and page-down
When paging through the lookup table, the preedit needs to be updated
to show the phrase selected by the cursor in the lookup table. Before
this patch, the preedit was not updated and when commiting the
character selected in the lookup table suddenly replaced the character
in the preedit. This differed from the behaviour when moving through
the lookup table with the arrow-up and arrow-down keys.
Using page-up and page-down or the arrow icons at the bottom
of the lookup table should also update the preedit just like it
is done when the arrow-up and arrow-down keys are used.